+++ CHANGES.rst
@@ -1,6 +1,8 @@
 **v1.1** July 2014 -- initial release
 
-* better reporting of differences between the old and updated bibtex entries
-* more options and improved documentation 
+* Added limited capability to check database entries against MathSciNet
+* Better reporting of differences between the old and updated bibtex entries
+* More options and improved documentation 
+* Optimised mrlookup searches and bibtex field extraction.
 
 **v1.0** July 2012 -- initial version.